Add some additional debug messages.
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@178124 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 89a9da4d6f51b82c20bc25fa3c83270220d8765b

+++ Rec/Rec/GlobalReco/src/ChargedProtoParticleMoni.h
@@ -101,7 +101,7 @@ private:
 private:
 
   /// Event count
-  unsigned long m_nEvts;
+  unsigned long long m_nEvts;
 
   /// Total number of tracks considered and selected
   TrackMap m_nTracks;
